# apps/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Feedback Flow API",
    "version": "1.7.0"
  },
  "paths": {
...
    "/api/feedback": {
      "post": {
        "summary": "Add feedback",
        "description": "Creates a new feedback record. Requires write permission.",
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "bearerAuth": [
              "write:api"
            ]
          },
          {
            "oauth2": [
              "write:api"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Feedback successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or missing required fields"
          }
        }
      }
    },
    "/api/publish": {
      "post": {
        "summary": "Record publication of feedback",
        "description": "Records the publication of feedback. Requires write permission.",
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "bearerAuth": [
              "write:api"
            ]
          },
          {
            "oauth2": [
              "write:api"
            ]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Publication successfully recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or missing required fields"
          }
        }
      }
    },
    "/api/publish/{id}": {
      "get": {
        "summary": "Get publication info",
        "description": "Returns information about a specific publication. Requires read permission.",
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "bearerAuth": [
              "read:api"
            ]
          },
          {
            "oauth2": [
              "read:api"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Purchase ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved publication info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Publication"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Publication not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FeedbackCreateRequest": {
        "type": "object",
        "required": [
          "date",
          "purchase",
          "feedback"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "purchase": {
            "type": "string"
          },
          "feedback": {
            "type": "string"
          }
        }
      },
      "PublishCreateRequest": {
        "type": "object",
        "required": [
          "date",
          "purchase",
          "screenshot"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "purchase": {
            "type": "string"
          },
          "screenshot": {
            "type": "string"
          }
        }
      },
      "Publication": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "purchase": {
            "type": "string"
          },
          "screenshot": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2": {
        "type": "oauth2",
        "description": "Auth0 OAuth2 Authorization Code Flow",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://sctg.eu.auth0.com/authorize",
            "tokenUrl": "https://sctg.eu.auth0.com/oauth/token",
            "scopes": {
              "read:api": "Read Api",
              "write:api": "Write Api",
              "admin:api": "Admin Api",
              "backup:api": "Backup Api",
              "search:api": "Search Api",
              "auth0:admin:api": "Auth0 Admin Api",
              "openid": "Openid",
              "profile": "Profile",
              "email": "Email"
            }
          }
        }
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "oauth2": []
    }
  ]
}